move view components into ProfileDetails
